Added 2016/13.

+++ 2017/5.h
@@ -11,10 +11,10 @@ struct solve<5> {
 	}
 	
 	template<typename F>
-	int simulate(std::vector<int> program, F&& f) const {
+	int simulate(std::vector<short> program, F&& f) const {
 		int cnt = 0;
-		for (int ip = 0; ip >= 0 && ip < program.size(); ++cnt) {
-			int next = ip + program[ip];
+		for (short ip = 0; ip >= 0 && ip < program.size(); ++cnt) {
+			short next = ip + program[ip];
 			program[ip] = f(program[ip]);
 			ip = next;
 		}
@@ -23,11 +23,11 @@ struct solve<5> {
 	
 	const std::string input = "2017/5.input";
 	auto operator()() const {
-		auto program = read_file_linewise(input, [](const auto& s){ return std::stoi(s); });
+		auto program = read_file_linewise(input, [](const auto& s){ return static_cast<short>(std::stoi(s)); });
 		
 		return std::make_pair(
-			simulate(program, [](int i){ return i+1; }),
-			simulate(program, [](int i){ return (i >= 3 ? i-1 : i+1); })
+			simulate(program, [](short i){ return i+1; }),
+			simulate(program, [](short i){ return (i >= 3 ? i-1 : i+1); })
 		);
 	}
 };
